fix: db today_count 오버플로우 나는 문제 해결

See merge request s11-webmobile1-sub2/S11P12D107!269

+++ backend/src/main/java/com/chestnut/backend/log/entity/StudyLog.java
@@ -37,8 +37,8 @@ public class StudyLog {
     @Column(columnDefinition = "date default (CURRENT_DATE)", nullable = false)
     private LocalDate studiedDay;
 
-    @Column(columnDefinition = "tinyint", nullable = false)
-    private byte todayCount;
+    @Column(columnDefinition = "int", nullable = false)
+    private int todayCount;
 
     @Column(columnDefinition = "tinyint(1) default 0", nullable = false)
     private boolean isPass;
